got rid of upload button
